Split common and __main__ into more modules
