+++ .cardboardlint.yml
@@ -7,7 +7,8 @@ linters:
   - import:
       packages: ['cardboardlint']
   - namespace:
-      filefilter: ['- */__init__.py', '- */test_*.py', '- *setup.py', '+ *.py', '+ *.pyx']
+      filefilter: ['- */__init__.py', '- */test_*.py', '- *setup.py',
+                   '- cardboardlint/linter_*.py', '+ *.py', '+ *.pyx']
   - pylint:
   - pycodestyle:
       config: .pycodestylerc
