+++ cardboardlint/linter_cppcheck.py
@@ -24,10 +24,12 @@ This test calls the cppcheck program, see http://cppcheck.sourceforge.net/.
 from xml.etree import ElementTree
 from typing import List
 
-from cardboardlint.common import run_command, Message, Linter
+from .linter import Linter
+from .message import Message
+from .utils import run_command
 
 
-__all__ = ['linter_cppcheck']
+__all__ = ['LINTER']
 
 
 DEFAULT_CONFIG = {
@@ -36,8 +38,8 @@ DEFAULT_CONFIG = {
 }
 
 
-def run_cppcheck(_config: dict, filenames: List[str], numproc: int = 1) -> List[Message]:
-    """Linter for cppcheck.
+def lint(_config: dict, filenames: List[str], numproc: int = 1) -> List[Message]:
+    """Lint with cppcheck.
 
     Parameters
     ----------
@@ -45,14 +47,14 @@ def run_cppcheck(_config: dict, filenames: List[str], numproc: int = 1) -> List[
         Dictionary that contains the configuration for the linter.
         _config is not used by this linter.
     filenames
-        A list of filenames to check
+        A list of filenames to check.
     numproc
         The number of processors to use.
 
     Returns
     -------
     messages
-        The list of messages generated by the external linter.
+        The list of messages generated by the linter.
 
     """
     # Get version
@@ -84,5 +86,4 @@ def run_cppcheck(_config: dict, filenames: List[str], numproc: int = 1) -> List[
     return messages
 
 
-# pylint: disable=invalid-name
-linter_cppcheck = Linter('cppcheck', run_cppcheck, DEFAULT_CONFIG, language='cpp')
+LINTER = Linter('cppcheck', lint, DEFAULT_CONFIG, language='cpp')
