+++ cardboardlint/linter_cpplint.py
@@ -23,10 +23,12 @@ This test calls the cpplint.py program, see https://github.com/google/styleguide
 
 from typing import List
 
-from cardboardlint.common import run_command, Message, Linter
+from .linter import Linter
+from .message import Message
+from .utils import run_command
 
 
-__all__ = ['linter_cpplint']
+__all__ = ['LINTER']
 
 
 DEFAULT_CONFIG = {
@@ -46,22 +48,22 @@ def _has_failed(_returncode, stdout, _stderr):
     return 'FATAL' in stdout
 
 
-def run_cpplint(config: dict, filenames: List[str], _numproc: int = 1) -> List[Message]:
-    """Linter for cpplint.
+def lint(config: dict, filenames: List[str], _numproc: int = 1) -> List[Message]:
+    """Lint with cpplint.
 
     Parameters
     ----------
     config
         Dictionary that contains the configuration for the linter.
     filenames
-        A list of filenames to check
+        A list of filenames to check.
     _numproc
         The number of processors to use.
 
     Returns
     -------
     messages
-        The list of messages generated by the external linter.
+        The list of messages generated by the linter.
 
     """
     messages = []
@@ -90,5 +92,4 @@ def run_cpplint(config: dict, filenames: List[str], _numproc: int = 1) -> List[M
     return messages
 
 
-# pylint: disable=invalid-name
-linter_cpplint = Linter('cpplint', run_cpplint, DEFAULT_CONFIG, language='cpp')
+LINTER = Linter('cpplint', lint, DEFAULT_CONFIG, language='cpp')
