+++ cardboardlint/linter_doxygen.py
@@ -24,10 +24,12 @@ This linter uses doxygen, see http://www.doxygen.org.
 import os
 from typing import List
 
-from cardboardlint.common import run_command, Message, Linter
+from .linter import Linter
+from .message import Message
+from .utils import run_command
 
 
-__all__ = ['linter_doxygen']
+__all__ = ['LINTER']
 
 
 DEFAULT_CONFIG = {
@@ -50,8 +52,8 @@ GENERATE_AUTOGEN_DEF   = NO
 """
 
 
-def run_doxygen(_config: dict, filenames: List[str], _numproc: int = 1) -> List[Message]:
-    """Linter using doxygen to find undocumented cpp.
+def lint(_config: dict, filenames: List[str], _numproc: int = 1) -> List[Message]:
+    """Lint with doxygen to find undocumented headers.
 
     Parameters
     ----------
@@ -59,14 +61,14 @@ def run_doxygen(_config: dict, filenames: List[str], _numproc: int = 1) -> List[
         Dictionary that contains the configuration for the linter.
         _config is not used by this linter.
     filenames
-        A list of filenames to check
+        A list of filenames to check.
     _numproc
         The number of processors to use.
 
     Returns
     -------
     messages
-        The list of messages generated by the external linter.
+        The list of messages generated by the linter.
 
     """
     command = ['doxygen', '--version']
@@ -96,5 +98,4 @@ def run_doxygen(_config: dict, filenames: List[str], _numproc: int = 1) -> List[
     return messages
 
 
-# pylint: disable=invalid-name
-linter_doxygen = Linter('doxygen', run_doxygen, DEFAULT_CONFIG, language='cpp')
+LINTER = Linter('doxygen', lint, DEFAULT_CONFIG, language='cpp')
