+++ cardboardlint/linter_flake8.py
@@ -23,10 +23,12 @@ This test calls the flake program, see http://flake8.pycqa.org
 
 from typing import List
 
-from cardboardlint.common import Message, run_command, Linter
+from .linter import Linter
+from .message import Message
+from .utils import run_command
 
 
-__all__ = ['linter_flake8']
+__all__ = ['LINTER']
 
 
 DEFAULT_CONFIG = {
@@ -42,22 +44,22 @@ def _has_failed(returncode, _stdout, _stderr):
     return not 0 <= returncode < 2
 
 
-def run_flake8(config: dict, filenames: List[str], numproc: int = 1) -> List[Message]:
-    """Linter for checking flake8 results.
+def lint(config: dict, filenames: List[str], numproc: int = 1) -> List[Message]:
+    """Lint with flake8.
 
     Parameters
     ----------
     config
         Dictionary that contains the configuration for the linter.
     filenames
-        A list of filenames to check
+        A list of filenames to check.
     numproc
         The number of processors to use.
 
     Returns
     -------
     messages
-        The list of messages generated by the external linter.
+        The list of messages generated by the linter.
 
     """
     # get flake8 version
@@ -79,5 +81,4 @@ def run_flake8(config: dict, filenames: List[str], numproc: int = 1) -> List[Mes
     return messages
 
 
-# pylint: disable=invalid-name
-linter_flake8 = Linter('flake8', run_flake8, DEFAULT_CONFIG, language='python')
+LINTER = Linter('flake8', lint, DEFAULT_CONFIG, language='python')
