+++ cardboardlint/linter_header.py
@@ -27,10 +27,11 @@ present, such as a shebang line or a mode line.
 import codecs
 from typing import List
 
-from cardboardlint.common import Message, Linter
+from .linter import Linter
+from .message import Message
 
 
-__all__ = ['linter_header']
+__all__ = ['LINTER']
 
 
 DEFAULT_CONFIG = {
@@ -48,22 +49,22 @@ DEFAULT_CONFIG = {
 }
 
 
-def run_header(config: dict, filenames: List[str], _numproc: int = 1) -> List[Message]:
-    """Linter for checking source file headers.
+def lint(config: dict, filenames: List[str], _numproc: int = 1) -> List[Message]:
+    """Lint source file headers.
 
     Parameters
     ----------
     config
         Dictionary that contains the configuration for the linter.
     filenames
-        A list of filenames to check
+        A list of filenames to check.
     _numproc
         The number of processors to use.
 
     Returns
     -------
     messages
-        The list of messages generated by the external linter.
+        The list of messages generated by the linter.
 
     """
     # Load the header file as a set of lines
@@ -119,5 +120,4 @@ def _check_file(filename: str, config: dict, header_lines: List[str],
                 header_counter += 1
 
 
-# pylint: disable=invalid-name
-linter_header = Linter('header', run_header, DEFAULT_CONFIG)
+LINTER = Linter('header', lint, DEFAULT_CONFIG)
