+++ cardboardlint/linter_import.py
@@ -30,10 +30,11 @@ This script counts the number of bad imports. The following is not allowed in a 
 import codecs
 from typing import List
 
-from cardboardlint.common import Message, Linter
+from .linter import Linter
+from .message import Message
 
 
-__all__ = ['linter_import']
+__all__ = ['LINTER']
 
 
 DEFAULT_CONFIG = {
@@ -44,22 +45,22 @@ DEFAULT_CONFIG = {
 }
 
 
-def run_import(config: dict, filenames: List[str], _numproc: int = 1) -> List[Message]:
-    """Linter for checking import statements.
+def lint(config: dict, filenames: List[str], _numproc: int = 1) -> List[Message]:
+    """Lint import statements.
 
     Parameters
     ----------
     config
         Dictionary that contains the configuration for the linter.
     filenames
-        A list of filenames to check
+        A list of filenames to check.
     numproc
         The number of processors to use.
 
     Returns
     -------
     _messages
-        The list of messages generated by the external linter.
+        The list of messages generated by the linter.
 
     """
     # Loop all python and cython files
@@ -97,5 +98,4 @@ def _check_file(filename: str, config: dict, messages: List[str]):
                     messages.append(Message(filename, lineno+1, None, text))
 
 
-# pylint: disable=invalid-name
-linter_import = Linter('import', run_import, DEFAULT_CONFIG, language='python')
+LINTER = Linter('import', lint, DEFAULT_CONFIG, language='python')
