+++ cardboardlint/linter_namespace.py
@@ -28,10 +28,11 @@ from typing import List
 import sys
 import importlib
 
-from cardboardlint.common import Message, Linter
+from .linter import Linter
+from .message import Message
 
 
-__all__ = ['linter_namespace']
+__all__ = ['LINTER']
 
 
 DEFAULT_CONFIG = {
@@ -42,22 +43,22 @@ DEFAULT_CONFIG = {
 }
 
 
-def run_namespace(config: dict, filenames: List[str], _numproc: int = 1) -> List[Message]:
-    """Linter for checking namespace Python namespace collisions.
+def lint(config: dict, filenames: List[str], _numproc: int = 1) -> List[Message]:
+    """Lint Python namespace collisions.
 
     Parameters
     ----------
     config
         Dictionary that contains the configuration for the linter.
     filenames
-        A list of filenames to check
+        A list of filenames to check.
     _numproc
         The number of processors to use.
 
     Returns
     -------
     messages
-        The list of messages generated by the external linter.
+        The list of messages generated by the linter.
 
     """
     # Make sure we test the source tree and not some locally installed copy of HORTON.
@@ -96,6 +97,4 @@ def run_namespace(config: dict, filenames: List[str], _numproc: int = 1) -> List
     return messages
 
 
-# pylint: disable=invalid-name
-linter_namespace = Linter('namespace', run_namespace, DEFAULT_CONFIG,
-                          style='dynamic', language='python')
+LINTER = Linter('namespace', lint, DEFAULT_CONFIG, style='dynamic', language='python')
