+++ cardboardlint/linter_pycodestyle.py
@@ -23,10 +23,12 @@ This test calls the flake program, see http://pycodestyle.pycqa.org
 
 from typing import List
 
-from cardboardlint.common import Message, run_command, Linter
+from .linter import Linter
+from .message import Message
+from .utils import run_command
 
 
-__all__ = ['linter_pycodestyle']
+__all__ = ['LINTER']
 
 
 DEFAULT_CONFIG = {
@@ -37,22 +39,22 @@ DEFAULT_CONFIG = {
 }
 
 
-def run_pycodestyle(config: dict, filenames: List[str], _numproc: int = 1) -> List[Message]:
-    """Linter for checking pycodestyle results.
+def lint(config: dict, filenames: List[str], _numproc: int = 1) -> List[Message]:
+    """Lint with pycodestyle.
 
     Parameters
     ----------
     config
         Dictionary that contains the configuration for the linter.
     filenames
-        A list of filenames to check
+        A list of filenames to check.
     _numproc
         The number of processors to use.
 
     Returns
     -------
     messages
-        The list of messages generated by the external linter.
+        The list of messages generated by the linter.
 
     """
     # get pycodestyle version
@@ -78,5 +80,4 @@ def run_pycodestyle(config: dict, filenames: List[str], _numproc: int = 1) -> Li
     return messages
 
 
-# pylint: disable=invalid-name
-linter_pycodestyle = Linter('pycodestyle', run_pycodestyle, DEFAULT_CONFIG, language='python')
+LINTER = Linter('pycodestyle', lint, DEFAULT_CONFIG, language='python')
