+++ cardboardlint/linter_pydocstyle.py
@@ -23,10 +23,12 @@ This test calls the flake program, see http://pydocstyle.pycqa.org
 
 from typing import List
 
-from cardboardlint.common import Message, run_command, Linter
+from .linter import Linter
+from .message import Message
+from .utils import run_command
 
 
-__all__ = ['linter_pydocstyle']
+__all__ = ['LINTER']
 
 
 DEFAULT_CONFIG = {
@@ -37,7 +39,7 @@ DEFAULT_CONFIG = {
 }
 
 
-def run_pydocstyle(config: dict, filenames: List[str], _numproc: int = 1) -> List[Message]:
+def lint(config: dict, filenames: List[str], _numproc: int = 1) -> List[Message]:
     """Linter for checking pydocstyle results.
 
     Parameters
@@ -45,14 +47,14 @@ def run_pydocstyle(config: dict, filenames: List[str], _numproc: int = 1) -> Lis
     config
         Dictionary that contains the configuration for the linter.
     filenames
-        A list of filenames to check
+        A list of filenames to check.
     _numproc
         The number of processors to use.
 
     Returns
     -------
     messages
-        The list of messages generated by the external linter.
+        The list of messages generated by the linter.
 
     """
     # get pydocstyle version
@@ -85,5 +87,4 @@ def run_pydocstyle(config: dict, filenames: List[str], _numproc: int = 1) -> Lis
     return messages
 
 
-# pylint: disable=invalid-name
-linter_pydocstyle = Linter('pydocstyle', run_pydocstyle, DEFAULT_CONFIG, language='python')
+LINTER = Linter('pydocstyle', lint, DEFAULT_CONFIG, language='python')
