+++ cardboardlint/linter_pylint.py
@@ -24,10 +24,12 @@ This test calls the pylint program, see http://docs.pylint.org/index.html.
 import json
 from typing import List
 
-from cardboardlint.common import Message, run_command, Linter
+from .linter import Linter
+from .message import Message
+from .utils import run_command
 
 
-__all__ = ['linter_pylint']
+__all__ = ['LINTER']
 
 
 DEFAULT_CONFIG = {
@@ -38,22 +40,22 @@ DEFAULT_CONFIG = {
 }
 
 
-def run_pylint(config: dict, filenames: List[str], numproc: int = 1) -> List[Message]:
-    """Linter for checking pylint results.
+def lint(config: dict, filenames: List[str], numproc: int = 1) -> List[Message]:
+    """Lint with pylint.
 
     Parameters
     ----------
     config
         Dictionary that contains the configuration for the linter.
     filenames
-        A list of filenames to check
+        A list of filenames to check.
     numproc
         The number of processors to use.
 
     Returns
     -------
     messages
-        The list of messages generated by the external linter.
+        The list of messages generated by the linter.
 
     """
     # get Pylint version
@@ -85,5 +87,4 @@ def run_pylint(config: dict, filenames: List[str], numproc: int = 1) -> List[Mes
 
 # Pylint should be considered dynamic, which is in practice only true for projects with
 # extension modules.
-# pylint: disable=invalid-name
-linter_pylint = Linter('pylint', run_pylint, DEFAULT_CONFIG, style='dynamic', language='python')
+LINTER = Linter('pylint', lint, DEFAULT_CONFIG, style='dynamic', language='python')
