+++ cardboardlint/linter_yamllint.py
@@ -23,10 +23,12 @@ This test calls the flake program, see https://yamllint.readthedocs.io/en/latest
 
 from typing import List
 
-from cardboardlint.common import Message, run_command, Linter
+from .linter import Linter
+from .message import Message
+from .utils import run_command
 
 
-__all__ = ['linter_yamllint']
+__all__ = ['LINTER']
 
 
 DEFAULT_CONFIG = {
@@ -37,22 +39,22 @@ DEFAULT_CONFIG = {
 }
 
 
-def run_yamllint(config: dict, filenames: List[str], _numproc: int = 1) -> List[Message]:
-    """Linter for checking yamllint results.
+def lint(config: dict, filenames: List[str], _numproc: int = 1) -> List[Message]:
+    """Lint with yamllint.
 
     Parameters
     ----------
     config
         Dictionary that contains the configuration for the linter.
     filenames
-        A list of filenames to check
+        A list of filenames to check.
     _numproc
         The number of processors to use.
 
     Returns
     -------
     messages
-        The list of messages generated by the external linter.
+        The list of messages generated by the linter.
 
     """
     # get yamllint version
@@ -78,5 +80,4 @@ def run_yamllint(config: dict, filenames: List[str], _numproc: int = 1) -> List[
     return messages
 
 
-# pylint: disable=invalid-name
-linter_yamllint = Linter('yamllint', run_yamllint, DEFAULT_CONFIG, language='yaml')
+LINTER = Linter('yamllint', lint, DEFAULT_CONFIG, language='yaml')
